requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
opencv-python-headless==4.10.0.84
numpy>=1.24.0
Pillow==10.4.0
python-multipart==0.0.12
scikit-image==0.24.0
